Restructured Project
Created a TextList class which is responsible for holding the needed values and execute the sorting and filtering operations.
Renamed Data class to XmlReader because now it is only responsible for reading the given xml file.
Removed hard coded string values from main method. They are now constants in the InputConfig class.
Similarly The XmlReader class has the needed configurations in XmlConfig class.
Moved file checking to a separate method inside XmlReader.
Moved sortByValue method from TextList to a new MapUtils class, as it is a generic method.
Simplified and shortened most of the methods in TestList.
Moved the argument checking and feedback printing from main method to separate methods.
Created a private FilterAndOrder class in Main class for holding the values for sorting and filtering.